Fix designate config
The changes in [0] sadly were almost, but not quite, completely broken.
They introduced config options that do not exist in designate and also
made secrets being saved into the node config.

Fix all this, also clean up the remainder of a workaround that is no
longer needed.

This was partially fixed for stable/queens in [1].

backport: stable/queens

[0] If4e185eb58a525e22971c7a8b876274c353db847
[1] https://review.opendev.org/606094

Change-Id: I5ccb4fa5d8f3a003c4dd1b07e1c9ab2a99246dd6